+++ momadm_benchmarks/envs/crazyrl/surround/surround.py
@@ -200,7 +200,7 @@ class Surround(CrazyRLBaseParallelEnv):
                 reward_close_to_target = -10
 
             if self.multi_obj:
-                reward[agent] = np.array([reward_close_to_target, reward_far_from_other_agents])
+                reward[agent] = np.array([reward_close_to_target, reward_far_from_other_agents], dtype=np.float32)
             else:
                 # MO reward linearly combined using hardcoded weights
                 reward[agent] = 0.8 * reward_close_to_target + 0.2 * reward_far_from_other_agents
